mine-dashboard/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
requests==2.31.0
pytz==2023.3
paho-mqtt==1.6.1
influxdb-client==1.38.0
psycopg2-binary==2.9.9
pyserial==3.5 